+++ fft/compare.c
@@ -12,10 +12,10 @@ extern int failed;
 int
 compare_complex_results (const char *name_a, const gsl_complex a[],
 			 const char *name_b, const gsl_complex b[],
-			 unsigned int n,
+			 size_t n,
 			 const double allowed_ticks)
 {
-  unsigned int i;
+  size_t i;
   double ticks, max_ticks = 0;
   double dr, di;
   const char *flag;
