+++ fft/compare.h
@@ -3,5 +3,5 @@
 int
   compare_complex_results (const char *name_a, const gsl_complex a[],
 			   const char *name_b, const gsl_complex b[],
-			   unsigned int n,
+			   size_t n,
 			   const double allowed_ticks);
