+++ fft/complex.c
@@ -9,7 +9,7 @@
 
 int
 gsl_fft_complex_forward (gsl_complex data[],
-			 const unsigned int n,
+			 const size_t n,
 			 const gsl_fft_complex_wavetable * wavetable)
 {
   gsl_fft_direction sign = forward;
@@ -19,7 +19,7 @@ gsl_fft_complex_forward (gsl_complex data[],
 
 int
 gsl_fft_complex_backward (gsl_complex data[],
-			  const unsigned int n,
+			  const size_t n,
 			  const gsl_fft_complex_wavetable * wavetable)
 {
   gsl_fft_direction sign = backward;
@@ -29,7 +29,7 @@ gsl_fft_complex_backward (gsl_complex data[],
 
 int
 gsl_fft_complex_inverse (gsl_complex data[],
-			 const unsigned int n,
+			 const size_t n,
 			 const gsl_fft_complex_wavetable * wavetable)
 {
   gsl_fft_direction sign = backward;
@@ -44,7 +44,7 @@ gsl_fft_complex_inverse (gsl_complex data[],
 
   {
     const double norm = 1.0 / n;
-    unsigned int i;
+    size_t i;
     for (i = 0; i < n; i++)
       {
 	data[i].real *= norm;
@@ -56,21 +56,21 @@ gsl_fft_complex_inverse (gsl_complex data[],
 
 int
 gsl_fft_complex (gsl_complex data[],
-		 const unsigned int n,
+		 const size_t n,
 		 const gsl_fft_complex_wavetable * wavetable,
 		 const gsl_fft_direction sign)
 {
 
-  const unsigned int nf = wavetable->nf;
+  const size_t nf = wavetable->nf;
 
-  unsigned int i;
+  size_t i;
 
-  unsigned int q, product = 1;
+  size_t q, product = 1;
 
   gsl_complex *scratch = wavetable->scratch;
   gsl_complex *twiddle1, *twiddle2, *twiddle3, *twiddle4, *twiddle5, *twiddle6;
 
-  unsigned int state = 0;
+  size_t state = 0;
   gsl_complex *from = data;
   gsl_complex *to = scratch;
 
@@ -92,7 +92,7 @@ gsl_fft_complex (gsl_complex data[],
 
   for (i = 0; i < nf; i++)
     {
-      const unsigned int factor = wavetable->factor[i];
+      const size_t factor = wavetable->factor[i];
       product *= factor;
       q = n / product;
