+++ fft/complex_pass_7.c
@@ -8,8 +8,8 @@ int
 gsl_fft_complex_pass_7 (const gsl_complex from[],
 			gsl_complex to[],
 			const gsl_fft_direction sign,
-			const unsigned int product,
-			const unsigned int n,
+			const size_t product,
+			const size_t n,
 			const gsl_complex twiddle1[],
 			const gsl_complex twiddle2[],
 			const gsl_complex twiddle3[],
@@ -18,14 +18,14 @@ gsl_fft_complex_pass_7 (const gsl_complex from[],
 			const gsl_complex twiddle6[])
 {
 
-  unsigned int i = 0, j = 0;
-  unsigned int k, k1;
+  size_t i = 0, j = 0;
+  size_t k, k1;
 
-  const unsigned int factor = 7;
-  const unsigned int m = n / factor;
-  const unsigned int q = n / product;
-  const unsigned int product_1 = product / factor;
-  const unsigned int jump = (factor - 1) * product_1;
+  const size_t factor = 7;
+  const size_t m = n / factor;
+  const size_t q = n / product;
+  const size_t product_1 = product / factor;
+  const size_t jump = (factor - 1) * product_1;
 
   const double c1 = cos(1.0 * 2.0 * M_PI / 7.0) ;
   const double c2 = cos(2.0 * 2.0 * M_PI / 7.0) ;
@@ -96,13 +96,13 @@ gsl_fft_complex_pass_7 (const gsl_complex from[],
 	  gsl_complex z0, z1, z2, z3, z4, z5, z6;
 
 	  {
-	    const unsigned int from0 = i;
-	    const unsigned int from1 = from0 + m;
-	    const unsigned int from2 = from1 + m;
-	    const unsigned int from3 = from2 + m;
-	    const unsigned int from4 = from3 + m;
-	    const unsigned int from5 = from4 + m;
-	    const unsigned int from6 = from5 + m;
+	    const size_t from0 = i;
+	    const size_t from1 = from0 + m;
+	    const size_t from2 = from1 + m;
+	    const size_t from3 = from2 + m;
+	    const size_t from4 = from3 + m;
+	    const size_t from5 = from4 + m;
+	    const size_t from6 = from5 + m;
 
 	    z0 = from[from0];
 	    z1 = from[from1];
@@ -268,13 +268,13 @@ gsl_fft_complex_pass_7 (const gsl_complex from[],
 
 	  /* apply twiddle factors */
 
-	    const unsigned int to0 = j;
-	    const unsigned int to1 = to0 + product_1;
-	    const unsigned int to2 = to1 + product_1;
-	    const unsigned int to3 = to2 + product_1;
-	    const unsigned int to4 = to3 + product_1;
-	    const unsigned int to5 = to4 + product_1;
-	    const unsigned int to6 = to5 + product_1;
+	    const size_t to0 = j;
+	    const size_t to1 = to0 + product_1;
+	    const size_t to2 = to1 + product_1;
+	    const size_t to3 = to2 + product_1;
+	    const size_t to4 = to3 + product_1;
+	    const size_t to5 = to4 + product_1;
+	    const size_t to6 = to5 + product_1;
 
 	    /* to0 = 1 * x0 */
 	    to[to0].real = x0_real;
