+++ fft/gsl_dft_complex.h
@@ -1,6 +1,8 @@
 #ifndef _GSL_DFT_COMPLEX_H
 #define _GSL_DFT_COMPLEX_H
 
+#include <stddef.h>
+
 #include <gsl_math.h>
 
 #include <gsl_complex.h>
@@ -9,22 +11,22 @@
 int
   gsl_dft_complex_forward (const gsl_complex data[],
 			   gsl_complex result[],
-			   const unsigned int n);
+			   const size_t n);
 
 int
   gsl_dft_complex_backward (const gsl_complex data[],
 			    gsl_complex result[],
-			    const unsigned int n);
+			    const size_t n);
 
 int
   gsl_dft_complex_inverse (const gsl_complex data[],
 			   gsl_complex result[],
-			   const unsigned int n);
+			   const size_t n);
 
 int
   gsl_dft_complex (const gsl_complex data[],
 		   gsl_complex result[],
-		   const unsigned int n,
+		   const size_t n,
 		   const gsl_fft_direction sign);
