+++ fft/gsl_fft_halfcomplex.h
@@ -1,27 +1,29 @@
 #ifndef _GSL_FFT_HALFCOMPLEX_H
 #define _GSL_FFT_HALFCOMPLEX_H
 
+#include <stddef.h>
+
 #include <gsl_math.h>
 #include <gsl_complex.h>
 #include <gsl_fft.h>
 
 int
 gsl_fft_halfcomplex_radix2_backward (double data[],
-				      const unsigned int n) ;
+				      const size_t n) ;
 
 int
 gsl_fft_halfcomplex_radix2_inverse (double data[],
-				    const unsigned int n) ;
+				    const size_t n) ;
 
 int
 gsl_fft_halfcomplex_radix2 (double data[],
-			    const unsigned int n) ;
+			    const size_t n) ;
 
 typedef struct
   {
-    unsigned int n;
-    unsigned int nf;
-    unsigned int factor[64];
+    size_t n;
+    size_t nf;
+    size_t factor[64];
     gsl_complex *twiddle[64];
     gsl_complex *trig;
     gsl_complex *scratch;
@@ -30,29 +32,29 @@ gsl_fft_halfcomplex_wavetable;
 
 int
   gsl_fft_halfcomplex_backward (double *data,
-				 const unsigned int n,
+				 const size_t n,
 			   const gsl_fft_halfcomplex_wavetable * wavetable);
 
 int
   gsl_fft_halfcomplex_inverse (double *data,
-			       const unsigned int n,
+			       const size_t n,
 			   const gsl_fft_halfcomplex_wavetable * wavetable);
 
 int
   gsl_fft_halfcomplex (double *data,
-		       const unsigned int n,
+		       const size_t n,
 		       const gsl_fft_halfcomplex_wavetable * wavetable);
 
 int
-  gsl_fft_halfcomplex_init (unsigned int n,
+  gsl_fft_halfcomplex_init (size_t n,
 			    gsl_fft_halfcomplex_wavetable * wavetable);
 
 int
-  gsl_fft_halfcomplex_generate_wavetable (unsigned int n,
+  gsl_fft_halfcomplex_generate_wavetable (size_t n,
 				 gsl_fft_halfcomplex_wavetable * wavetable);
 
 int
-  gsl_fft_halfcomplex_wavetable_alloc (unsigned int n,
+  gsl_fft_halfcomplex_wavetable_alloc (size_t n,
 				 gsl_fft_halfcomplex_wavetable * wavetable);
 
 int
@@ -61,6 +63,6 @@ int
 int
   gsl_fft_halfcomplex_unpack (const double halfcomplex_coefficient[],
 			      gsl_complex complex_coefficient[],
-			      const unsigned int n);
+			      const size_t n);
 
 #endif /* _GSL_FFT_HALFCOMPLEX_H */
