+++ fft/halfcomplex_pass_2.c
@@ -7,15 +7,15 @@
 int
 gsl_fft_halfcomplex_pass_2 (const double from[],
 			    double to[],
-			    const unsigned int product,
-			    const unsigned int n,
+			    const size_t product,
+			    const size_t n,
 			    const gsl_complex twiddle[])
 {
 
-  unsigned int i, j, k, k1, jump;
-  unsigned int factor, q, m, product_1;
-  unsigned int from0, from1;
-  unsigned int to0, to1;
+  size_t i, j, k, k1, jump;
+  size_t factor, q, m, product_1;
+  size_t from0, from1;
+  size_t to0, to1;
   gsl_complex x0, x1;
   gsl_complex w;
   gsl_complex z0, z1;
